# frontend/packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xian API",
    "description": "AI Agent 红蓝对抗平台 HTTP API（PRD 3.1~3.9）",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/levels": {
      "get": {
        "tags": [
          "levels"
        ],
        "summary": "List Levels",
        "operationId": "list_levels_api_v1_levels_get",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LevelOut"
                  },
                  "title": "Response List Levels Api V1 Levels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/levels/progress": {
      "get": {
        "tags": [
          "levels"
        ],
        "summary": "My Progress",
        "operationId": "my_progress_api_v1_levels_progress_get",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LevelProgressOut"
                  },
                  "title": "Response My Progress Api V1 Levels Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/levels/{code}/start": {
      "post": {
        "tags": [
          "levels"
        ],
        "summary": "Start Level",
        "operationId": "start_level_api_v1_levels__code__start_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LevelProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/levels/{code}/hint": {
      "post": {
        "tags": [
          "levels"
        ],
        "summary": "Use Hint",
        "description": "三级提示：H1 方向 / H2 思路 / H3 近似 payload，分别扣 10/20/40 能量。",
        "operationId": "use_hint_api_v1_levels__code__hint_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HintUseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HintOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/levels/{code}/submit": {
      "post": {
        "tags": [
          "levels"
        ],
        "summary": "Submit Attempt",
        "description": "提交一次尝试，按 PRD 3.4.5.1 可量化条件判定通关。",
        "operationId": "submit_attempt_api_v1_levels__code__submit_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LevelProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/levels/{code}/hardening": {
      "get": {
        "tags": [
          "levels"
        ],
        "summary": "Hardening",
        "operationId": "hardening_api_v1_levels__code__hardening_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "blue",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Hardening Api V1 Levels  Code  Hardening Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Difficulty": {
        "type": "string",
        "enum": [
          "beginner",
          "easy",
          "low",
          "medium",
          "hard",
          "expert"
        ],
        "title": "Difficulty"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HintOut": {
        "properties": {
          "level": {
            "type": "string",
            "title": "Level"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "energy_cost": {
            "type": "integer",
            "title": "Energy Cost"
          },
          "energy_left": {
            "type": "integer",
            "title": "Energy Left"
          }
        },
        "type": "object",
        "required": [
          "level",
          "content",
          "energy_cost",
          "energy_left"
        ],
        "title": "HintOut"
      },
      "HintUseIn": {
        "properties": {
          "hint_level": {
            "type": "string",
            "pattern": "^H[123]$",
            "title": "Hint Level"
          }
        },
        "type": "object",
        "required": [
          "hint_level"
        ],
        "title": "HintUseIn"
      },
...
      "LevelOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scenario_code": {
            "type": "string",
            "title": "Scenario Code"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "pass_criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pass Criteria"
          },
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques"
          },
          "hints": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Hints"
          },
          "unlock_rule": {
            "type": "string",
            "title": "Unlock Rule"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty",
            "default": "easy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "scenario_code",
          "goal",
          "pass_criteria",
          "techniques",
          "hints",
          "unlock_rule"
        ],
        "title": "LevelOut"
      },
      "LevelProgressOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "level_id": {
            "type": "string",
            "title": "Level Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "time_used": {
            "type": "integer",
            "title": "Time Used"
          },
          "hints_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hints Used"
          },
          "energy_left": {
            "type": "integer",
            "title": "Energy Left"
          },
          "dimension_coverage": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Dimension Coverage"
          },
          "badge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badge"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "level_id",
          "status",
          "score",
          "time_used",
          "hints_used",
          "energy_left",
          "dimension_coverage"
        ],
        "title": "LevelProgressOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}